# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent Server",
    "description": "Web API for the Modular Agent Framework — streaming plan→execute→reply",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/agent/impersonate/chat": {
      "post": {
        "summary": "Impersonate Chat",
        "operationId": "impersonate_chat_api_v1_agent_impersonate_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImpersonateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpersonateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/impersonate/chat/stream": {
      "post": {
        "summary": "Impersonate Chat Stream",
        "operationId": "impersonate_chat_stream_api_v1_agent_impersonate_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImpersonateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/impersonate/reset": {
      "post": {
        "summary": "Impersonate Reset",
        "operationId": "impersonate_reset_api_v1_agent_impersonate_reset_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Session ID",
              "title": "Session Id"
            },
            "description": "Session ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/impersonate/regenerate": {
      "post": {
        "summary": "Impersonate Regenerate",
        "operationId": "impersonate_regenerate_api_v1_agent_impersonate_regenerate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImpersonateRegenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/impersonate/history": {
      "get": {
        "summary": "Impersonate History",
        "operationId": "impersonate_history_api_v1_agent_impersonate_history_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Session ID",
              "title": "Session Id"
            },
            "description": "Session ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImpersonateRegenerateRequest": {
        "properties": {
          "character": {
            "type": "string",
            "minLength": 1,
            "title": "Character"
          },
          "session_id": {
            "type": "string",
            "minLength": 1,
            "title": "Session Id"
          },
          "doc_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Id"
          }
        },
        "type": "object",
        "required": [
          "character",
          "session_id"
        ],
        "title": "ImpersonateRegenerateRequest"
      },
      "ImpersonateRequest": {
        "properties": {
          "character": {
            "type": "string",
            "minLength": 1,
            "title": "Character",
            "description": "Character name"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "title": "Message",
            "description": "User message to character"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Session ID"
          },
          "doc_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Id",
            "description": "Optional volume lock for retrieval"
          }
        },
        "type": "object",
        "required": [
          "character",
          "message"
        ],
        "title": "ImpersonateRequest"
      },
      "ImpersonateResponse": {
        "properties": {
          "reply": {
            "type": "string",
            "title": "Reply"
          },
          "character": {
            "type": "string",
            "title": "Character"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "citations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Citations"
          }
        },
        "type": "object",
        "required": [
          "reply",
          "character",
          "session_id"
        ],
        "title": "ImpersonateResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}